Refactor ExcInfo class for improved structure
