+++ vedro/core/exc_info/__init__.py
@@ -1,0 +1,3 @@
+from ._exc_info import ExcInfo
+
+__all__ = ("ExcInfo",)
